fix(server): leave a connection to an idle worker and stream files from disk
A worker whose render queue is full no longer accepts from the socket it
shares with other workers: the connection waits in the kernel's queue for a
worker with room, rather than being taken only to be answered 503 while a
sibling sits idle. A server with no sibling still takes it and answers 503
at once.

readyz answers 503 once the server has begun to stop, as its docs say.

A file is copied from disk to the connection as it is written, with its
length from the open file, instead of being read whole into memory for every
request; a HEAD request opens it and reads nothing.

lumen.site.json is refused under any name that opens it, such as another
case on a case-insensitive file system, compared by file identity.

+++ web/server/src/log.rs
@@ -82,7 +82,7 @@ pub struct Access<'a> {
     /// The status sent.
     pub status: u16,
     /// The body bytes sent.
-    pub bytes: usize,
+    pub bytes: u64,
     /// From the first byte of the request to the last byte of the response.
     pub took: Duration,
 }
